@@ -1,64 +1,37 @@
 {
-    "name": "website",
+    "name": "web",
     "version": "0.1.0",
     "private": true,
     "scripts": {
-        "dev": "next dev",
+        "dev": "next dev --turbopack",
         "build": "next build",
         "start": "next start",
         "lint": "next lint",
-        "sh:add": "bunx shadcn-ui@latest add"
+        "ui:add": "bunx shadcn@latest add"
     },
     "dependencies": {
-        "@clerk/elements": "^0.10.7",
-        "@clerk/nextjs": "^5.2.8",
-        "@clerk/themes": "^2.1.14",
-        "@hookform/resolvers": "^3.3.4",
-        "@radix-ui/react-avatar": "^1.0.4",
-        "@radix-ui/react-dialog": "^1.0.5",
-        "@radix-ui/react-dropdown-menu": "^2.0.6",
-        "@radix-ui/react-icons": "^1.3.0",
-        "@radix-ui/react-label": "^2.0.2",
-        "@radix-ui/react-popover": "^1.0.7",
-        "@radix-ui/react-progress": "^1.0.3",
-        "@radix-ui/react-scroll-area": "^1.0.5",
-        "@radix-ui/react-select": "^2.0.0",
-        "@radix-ui/react-separator": "^1.0.3",
-        "@radix-ui/react-slot": "^1.0.2",
-        "@radix-ui/react-tooltip": "^1.0.7",
-        "axios": "^1.6.7",
-        "class-variance-authority": "^0.7.0",
-        "clsx": "^2.1.0",
-        "cmdk": "^1.0.0",
-        "framer-motion": "^11.0.8",
-        "lucide-react": "^0.344.0",
-        "mini-svg-data-uri": "^1.4.4",
-        "ms": "^2.1.3",
-        "next": "14.2.9",
-        "next-themes": "^0.2.1",
-        "posthog-js": "^1.116.0",
-        "react": "^18",
-        "react-dom": "^18",
-        "react-hook-form": "^7.51.0",
-        "sonner": "^1.4.3",
-        "tailwind-merge": "^2.2.1",
-        "tailwindcss-animate": "^1.0.7",
-        "zod": "^3.22.4",
-        "zustand": "^4.5.2"
+        "@radix-ui/react-slot": "^1.1.1",
+        "class-variance-authority": "^0.7.1",
+        "clsx": "^2.1.1",
+        "framer-motion": "^11.18.0",
+        "lucide-react": "^0.471.1",
+        "next": "15.1.4",
+        "react": "^19.0.0",
+        "react-dom": "^19.0.0",
+        "tailwind-merge": "^2.6.0",
+        "tailwindcss-animate": "^1.0.7"
     },
     "devDependencies": {
-        "@svgr/webpack": "^8.1.0",
-        "@types/ms": "^0.7.34",
+        "@eslint/eslintrc": "^3",
         "@types/node": "^20",
-        "@types/react": "^18",
-        "@types/react-dom": "^18",
-        "autoprefixer": "^10.0.1",
-        "eslint": "^8",
-        "eslint-config-next": "14.1.1",
+        "@types/react": "^19",
+        "@types/react-dom": "^19",
+        "eslint": "^9",
+        "eslint-config-next": "15.1.4",
         "postcss": "^8",
-        "prettier-plugin-organize-imports": "^4.0.0",
-        "prettier-plugin-tailwindcss": "^0.5.11",
-        "tailwindcss": "^3.3.0",
+        "prettier-plugin-organize-imports": "^4.1.0",
+        "prettier-plugin-tailwindcss": "^0.6.9",
+        "tailwindcss": "^3.4.1",
         "typescript": "^5"
     }
 }
